refactor(metrics): split geo.metrics into xr_toolz.metrics taxonomy (F1.1)
Stand up the new ``xr_toolz.metrics`` package with the submodule layout
specified in docs/design/validation.md. The 11 Layer-0 functions and 8
Layer-1 operator classes that previously lived in ``xr_toolz.geo`` move
to ``metrics._src.pixel`` (pointwise) and ``metrics._src.spectral``
(PSD-based). Nine empty stubs land alongside (multiscale, structural,
probabilistic, distributional, masked, forecast, object, physical,
lagrangian) so the V1–V5 view PRs have a home to drop into.

No math changes. The old import paths
(``xr_toolz.geo.{rmse,mse,…,psd_score,find_intercept_1D}`` and
``xr_toolz.geo.operators.{MSE,RMSE,…,PSDScore}``) keep working for one
release via lazy ``__getattr__`` shims that emit a
``DeprecationWarning`` only when a moved name is actually accessed —
plain ``import xr_toolz.geo`` stays silent.

Closes #23.
