+++ src/xr_toolz/geo/__init__.py
@@ -3,14 +3,22 @@
 This submodule hosts everything that applies across Earth-science
 domains: coordinate validation, spatial/temporal subsetting, masks,
 regridding-adjacent (interpolation), detrending, encoders, spectral
-analysis, pixel + spectral metrics, extremes, binning, and CRS
-utilities.
+analysis, extremes, binning, and CRS utilities.
 
 The public API is re-exported from :mod:`xr_toolz.geo._src`. For direct
 access to a specific Layer-0 module, import e.g.
 ``xr_toolz.geo._src.detrend``. Layer-1 ``Operator`` wrappers live in
 :mod:`xr_toolz.geo.operators`.
+
+Evaluation metrics (``mse``, ``rmse``, …, ``psd_score``, ``find_intercept_1D``)
+moved to :mod:`xr_toolz.metrics`. They remain importable from this
+module for one release with a :class:`DeprecationWarning`.
 """
+
+from __future__ import annotations
+
+import warnings
+from typing import Any
 
 from xr_toolz.geo._src.crs import (
     assign_crs,
@@ -73,19 +81,6 @@ from xr_toolz.geo._src.masks import (
     add_ocean_mask,
     apply_mask,
 )
-from xr_toolz.geo._src.metrics import (
-    bias,
-    correlation,
-    find_intercept_1D,
-    mae,
-    mse,
-    nrmse,
-    psd_error,
-    psd_score,
-    r2_score,
-    resolved_scale,
-    rmse,
-)
 from xr_toolz.geo._src.subset import (
     select_variables,
     subset_bbox,
@@ -99,6 +94,39 @@ from xr_toolz.geo._src.validation import (
 )
 
 
+# Names moved to xr_toolz.metrics — kept importable for one release with
+# a deprecation warning fired only on actual access (PEP 562).
+_DEPRECATED_METRICS = {
+    "bias": "xr_toolz.metrics._src.pixel",
+    "correlation": "xr_toolz.metrics._src.pixel",
+    "mae": "xr_toolz.metrics._src.pixel",
+    "mse": "xr_toolz.metrics._src.pixel",
+    "nrmse": "xr_toolz.metrics._src.pixel",
+    "r2_score": "xr_toolz.metrics._src.pixel",
+    "rmse": "xr_toolz.metrics._src.pixel",
+    "find_intercept_1D": "xr_toolz.metrics._src.spectral",
+    "psd_error": "xr_toolz.metrics._src.spectral",
+    "psd_score": "xr_toolz.metrics._src.spectral",
+    "resolved_scale": "xr_toolz.metrics._src.spectral",
+}
+
+
+def __getattr__(name: str) -> Any:
+    if name in _DEPRECATED_METRICS:
+        from importlib import import_module
+
+        warnings.warn(
+            f"xr_toolz.geo.{name} is deprecated; "
+            f"import from xr_toolz.metrics instead. "
+            f"This re-export will be removed in the next minor release.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        module = import_module(_DEPRECATED_METRICS[name])
+        return getattr(module, name)
+    raise AttributeError(f"module 'xr_toolz.geo' has no attribute {name!r}")
+
+
 __all__ = [
     "Grid",
     "Period",
@@ -109,7 +137,6 @@ __all__ = [
     "add_ocean_mask",
     "apply_mask",
     "assign_crs",
-    "bias",
     "bin_2d",
     "block_maxima",
     "block_minima",
@@ -120,14 +147,12 @@ __all__ = [
     "calculate_climatology_season",
     "calculate_climatology_smoothed",
     "coarsen",
-    "correlation",
     "cyclical_encode",
     "encode_time_cyclical",
     "encode_time_ordinal",
     "fillnan_rbf",
     "fillnan_spatial",
     "fillnan_temporal",
-    "find_intercept_1D",
     "fourier_features",
     "get_crs",
     "histogram_2d",
@@ -136,26 +161,18 @@ __all__ = [
     "lon_180_to_360",
     "lon_360_to_180",
     "lonlat_to_xy",
-    "mae",
-    "mse",
-    "nrmse",
     "points_to_grid",
     "positional_encoding",
     "pot_exceedances",
     "pot_threshold",
     "pp_counts",
     "pp_stats",
-    "psd_error",
-    "psd_score",
-    "r2_score",
     "random_fourier_features",
     "refine",
     "remove_climatology",
     "rename_coords",
     "reproject",
     "resample_time",
-    "resolved_scale",
-    "rmse",
     "select_variables",
     "subset_bbox",
     "subset_time",
