+++ src/xr_toolz/metrics/_src/masked.py
@@ -1,0 +1,6 @@
+"""Stub: lands with view V2 (Epic).
+
+This module is intentionally empty. It will be populated by the V2
+view epic and is shipped now so downstream PRs can import its path
+without depending on a package layout change.
+"""
