+++ src/xr_toolz/metrics/_src/multiscale.py
@@ -1,0 +1,6 @@
+"""Stub: lands with view V1 (Epic).
+
+This module is intentionally empty. It will be populated by the V1
+view epic and is shipped now so downstream PRs can import its path
+without depending on a package layout change.
+"""
